moved to int too much headache with long conversion

+++ src/main/java/com/engine/Dto/RequestDto.java
@@ -3,7 +3,7 @@ package com.engine.Dto;
 import java.util.List;
 
 public record RequestDto(
-        long maxMargin,
+        int maxMargin,
         List<CandidateTradeDto> candidateTrades
 ) {
 }
